simplify implementation
